Added exception catcher which reboots 750M
Previously, if the 750M initialization failed, no instance would have been created. Since "reboot" is a bound method, the user wouldn't have an opportunity to try rebooting the controller.

Now, the initialization script will rebooting the controller once if it has trouble initializing. "reboot" is still a bound method. Hopefully this is a sufficient solution.