Changed annoying behavior of egg5110 'measure' function
When I coded up the egg5110 measurement function, I added the
'feature' of querying the sensitivity on each call, so that
the number returned was an actual number of volts or millivolts.
It also returned a string like 'V', 'mV', whatever.

This turned out to be annoying because I hardly ever care.
The property objects now only return a number, representing
the fraction of full-scale. To get a real number, just
multiply this by the sensitivity.

The old behavior is available by calling the 'measure' method
directly, and using unit=True as a keyword argument.