Added support for MM3000 stage controller, delay stage shortcuts
Major structural changes to support the MM3000 with a minimum of excess code. Shared functionality with the ESP300 is implemented in the shared base class newport_stage.

I also added a mixin for delay stages used specifically to delay a beam. This enables stages to be moved easily in units of picoseconds.