Added on/off switch for MM3000
Added the 'on' property for the MM3000. A bit superfluous since the motors are turned on whenever a move command happens.

The short newport stage doesn't seem to play nice with the ESP300, even with the new shorty_stage class. behavior of the thorlabs actuator is also weird. I should either debug this behavior (possibly by setting home search speed to 10% of maximum) or add unit functionality to the MM3000.

the calibration functions I added for the ESP300 a few commits back aren't supported by the MM3000 2.2 firmware. I could just implement those conversions in python for complete parity between the controllers.