modified prologix for per-instrument read-after-write
The prologix read-after-write feature (on by default, switched
using '++auto') is useful for most instruments - it reads their
response over GPIB and immediately puts that in the serial read
buffer. But sometimes you want to turn it off.

This modification to the prologix driver turns on read-after-write
by default, but enables you to leave it off for certain instruments
that behave strangely with this feature (e.g., the new focus 6300
laser diode controller).

There should be no drawback to turning this feature off by default
(except slightly slower response times), but I want to test that out
with a few instruments first.