Decreased default prologix instrument wait time from 0.5sec to 0.1sec
Communicating with instruments that don't support read-after-write takes twice as long (due to the extra ++read command). In the case of the new focus laser this was frustrating me because the default lag time after each write is 0.5 seconds.

I think 0.1 seconds is a more reasonable default. Most instruments should not take longer than that to render a response over GPIB. If they do, pass the old 0.5 value instead of this new value.

There are probably some cases where this extra time hasn't been wasted - scans, for example, are probably more accurate due to the built-in 0.5 second sleeping (on top of whatever sleeping is implemented in the loop). I may have to up the in-loop sleeping to get the same accuracy as before, but I haven't tested this.