Added support for Burleigh WA-1000 wavemeter over RS232
Bare-bones support for the wavemeter.

No query controls have been written - the wavemeter writes its results continually into the buffer, and this code just reads it.

When invoked as a script, this displays a forever-updating plot of the wavelength over time. Neat application of the new live plotter functionality in the last commit.

To do: wl attribute isn't necessarily wavelength, it may also be GHz or wavenumber. For now I'm just using the front panel control to select what I want, but it'd be nice to implement python methods for this.