Added wanglib package description
Begin code organizational push. Writing the package description revealed to me that a lot of this library is seemingly cobbled together out of junk. I'd like to streamline it before I offer it up for my labmates to use. util.py is the worst, but before I change that I need to cut out a bunch of dependencies from the various instruments.