Removed all dependencies on util.Gpib from instruments
GPIB instruments no longer know about their 'default' GPIB address. That now has to be made explicit when instantiating any instrument.

Previously, it was possible to instantiate an Agilent RF generator (for example) with no arguments, and wanglib would automatically look for it on address 18 - the factory default. This worked because util.py would try importing various GPIB modules (like pyVISA or Gpib, the linux-gpib python driver) until it found one, then monkey-patched it to behave roughly like pyVISA's visa.instrument object, and used that for the bus object.

This would be great if it worked, but it was always kind of half-assed, and I never implemented anything to search for prologix controllers because that would have been hard. Ever since my internal GPIB card died none of my experimental routines have depended on this functionality at all.

Default GPIB addresses are also bad because they make implicit an important experimental detail that should really be explicitly coded in the python script that defines the experimental setup. This should bring wanglib closer to PEP 20 compliance.

The atrocious module-searcher in util.py is still present, but it'll be gone soon.

Also added general documentation for the instruments package.