Added facility for auto-appending serial termination characters.
The ESP300 GPIB interface is dead, so I'm switching over to RS232. A
wanglib.util.Serial object can be a drop-in replacement for
wanglib.prologix.instrument, so long as it is extended to affix its own
line terminators. This is similar to the PyVISA SerialInstrument
functionality.

Other RS232 instruments in the library (the two spectrometers)
instantiate their own connections, but I didn't implement that here. I'm
torn on whether this is the best approach - wanglib should know
how specific instruments behave (command syntax, flow control, etc.) but
stop short of making assumptions about the setup (which port the
instrument is to be found at).

I need to come up with an elegant way for a user to specify that an
instrument is at /dev/ttyWhatever, without also requiring that he or she
know its preferred baud rate, flow control setting, and command
terminators.