Changed behavior of sensitivity and measurement functions
I've decided it makes the most sense to have compatibility between the
two brands of lockins, even if it means double-queries of the EG&G 5110.
This works toward that goal.

Now, lock-in measurements (li.x, li.r and friends) will always return a
figure in volts. Same for sensitivity. To recover the old behavior, you
can do a direct call to li.measure or li.get_sensitivity with the `unit`
kwarg.