Fixed bug: applying a mask to a line would modify the original data.
The arrays returned by get_data are references to the original data
passed to the plot command. Hence, it is very important to not modify
them in-place.

>>> from pylab import *
>>> from wanglib.pylab_extensions import apply_mask

>>> x = linspace(0,1,100) * 2 * pi
>>> y = sin(x)

>>> line, = plot(x,y)
>>> Y = line.get_ydata()

>>> # this fixes bug
>>> Y = Y.copy()

>>> for i,pt in enumerate(Y):
>>>     Y[i] = 2 * pt
>>>
>>> line.set_ydata(Y)

>>> plot(x,y)

>>> show()

This kind of works the other way. Plotting y, then modifying y, and
calling get_ydata will return the modified y. However, the line does
not reflect the changes made. I'm not sure why that is.