prologix factory functions
These create a module-level registry of prologix controllers, so if one
is instantiated for an address which already exists, the existing one is
returned instead of making a new one. This fixes a bug related to
caching of the _addr variable.

There's still an issue of the equivalence of certain addresses (e.g.,
equivalence of Serial(0) and Serial('COM1'), etc. Fix that later