Plotgen across multiple axes
You can now provide a tuple of axes that is the same length as the
number of lines you expect, and have the lines plotted on multiple axes.

Fixed bug with the previous edit: **kwargs is now passed on again

Removed feature: function does not return anything anymore. Use gll,
etc, to retrieve the data.